# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "hostingaffe",
    "description": "The HTTP surface of one hostingaffe instance. Every instance is at its own address; everything but GET /version takes `Authorization: Bearer <token>`.",
    "version": "0.0.0-dev"
  },
  "paths": {
...
    "/projects/{key}/releases": {
      "get": {
        "tags": [
          "hostingaffe"
        ],
        "summary": "Every release of the project: the open one first, then published ones newest first.",
        "operationId": "ListReleases",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReleaseSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/projects/{key}/releases/{name}": {
      "get": {
        "tags": [
          "hostingaffe"
        ],
        "summary": "One release with the issues that shipped in it.",
        "operationId": "ReadRelease",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Release"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "hostingaffe"
        ],
        "summary": "Annotate the release notes, and correct the name of the newest publication. A field left out is left alone.",
        "operationId": "ChangeRelease",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeReleaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Release"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{key}/releases/publish": {
      "post": {
        "tags": [
          "hostingaffe"
        ],
        "summary": "Name and freeze the open release, then create the next open one.",
        "operationId": "PublishRelease",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishReleaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Release"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{key}/releases/{name}/retract": {
      "post": {
        "tags": [
          "hostingaffe"
        ],
        "summary": "Take the newest publication back: the release is the open one again and the empty open release goes. Refused once another followed it.",
        "operationId": "RetractRelease",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Release"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{key}/releases/{name}/issues/{issue}": {
      "put": {
        "tags": [
          "hostingaffe"
        ],
        "summary": "Put an issue into the open release by hand. A published release is a record and takes none.",
        "operationId": "AddIssueToRelease",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "issue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Release"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "hostingaffe"
        ],
        "summary": "Take an issue out of the open release: it has not shipped yet and does not belong.",
        "operationId": "RemoveIssueFromRelease",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "issue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Release"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/admin/smtp": {
      "get": {
        "tags": [
          "hostingaffe"
        ],
        "summary": "Whether transactional email is configured, without credentials.",
        "operationId": "ReadSmtpStatus",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmtpStatus"
                }
              }
            }
          }
        }
      }
    },
    "/admin/smtp/test": {
      "post": {
        "tags": [
          "hostingaffe"
        ],
        "summary": "Send a transactional-email test message. Administrators only.",
        "operationId": "SendTestEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTestEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BlockerRef": {
        "required": [
          "key",
          "open"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": [
              "null",
              "string"
            ]
          },
          "open": {
            "type": "boolean"
          }
        }
      },
...
      "ChangeReleaseRequest": {
        "required": [
          "name",
          "description"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "Claim": {
        "required": [
          "holder",
          "since",
          "expires_at"
        ],
        "type": "object",
        "properties": {
          "holder": {
            "$ref": "#/components/schemas/IdentityRef"
          },
          "since": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "IdentityKind": {
        "enum": [
          "user",
          "agent"
        ],
        "type": "string"
      },
      "IdentityRef": {
        "required": [
          "id",
          "kind",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/IdentityKind"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "IssueStatus": {
        "enum": [
          "backlog",
          "todo",
          "in_progress",
          "review",
          "done",
          "canceled"
        ],
        "type": "string"
      },
      "IssueSummary": {
        "required": [
          "key",
          "project",
          "title",
          "status",
          "ready",
          "priority",
          "labels",
          "epic",
          "parent",
          "release",
          "assignee",
          "claim",
          "blocked_by",
          "open_questions",
          "open_blockers",
          "open_sub_issues",
          "created_at",
          "updated_at",
          "closed_at",
          "deleted_at",
          "deleted_by"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "ready": {
            "type": "boolean"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "epic": {
            "type": [
              "null",
              "string"
            ]
          },
          "parent": {
            "type": [
              "null",
              "string"
            ]
          },
          "release": {
            "type": [
              "null",
              "string"
            ]
          },
          "assignee": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IdentityRef"
              }
            ]
          },
          "claim": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Claim"
              }
            ]
          },
          "blocked_by": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockerRef"
            }
          },
          "open_questions": {
            "type": "integer",
            "format": "int32"
          },
          "open_blockers": {
            "type": "integer",
            "format": "int32"
          },
          "open_sub_issues": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "closed_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "deleted_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "deleted_by": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IdentityRef"
              }
            ]
          }
        }
      },
...
      "Priority": {
        "maximum": 4,
        "minimum": 0,
        "type": "integer"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PublishReleaseRequest": {
        "required": [
          "name",
          "description"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "Release": {
        "required": [
          "name",
          "status",
          "description",
          "published_at",
          "published_by",
          "issues"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ReleaseStatus"
          },
          "description": {
            "type": "string"
          },
          "published_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "published_by": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IdentityRef"
              }
            ]
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueSummary"
            }
          }
        }
      },
      "ReleaseStatus": {
        "enum": [
          "open",
          "published"
        ],
        "type": "string"
      },
      "ReleaseSummary": {
        "required": [
          "name",
          "status",
          "description",
          "published_at",
          "published_by",
          "issues"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ReleaseStatus"
          },
          "description": {
            "type": "string"
          },
          "published_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "published_by": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IdentityRef"
              }
            ]
          },
          "issues": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "SendTestEmailRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SmtpStatus": {
        "required": [
          "configured",
          "host",
          "port",
          "security",
          "sender"
        ],
        "type": "object",
        "properties": {
          "configured": {
            "type": "boolean"
          },
          "host": {
            "type": [
              "null",
              "string"
            ]
          },
          "port": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "security": {
            "type": [
              "null",
              "string"
            ]
          },
          "sender": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}